Use razor component to encapsulate UI logic
