+++ BTCPayServer/Views/PaymentRequest/GetPaymentRequests.cshtml
@@ -91,7 +91,7 @@
                     </tbody>
                 </table>
 
-                <partial name="_TableFooterPager" />
+                <vc:pager view-model="Model"></vc:pager>
             </div>
         </div>
     </div>
