+++ BTCPayServer/_ViewImports.cshtml
@@ -10,3 +10,4 @@
 @using Microsoft.AspNetCore.Routing;
 @inject BTCPayServer.Services.Safe Safe
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+@addTagHelper *, BTCPayServer
